% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions.R
\docType{data}
\name{mMSNBC}
\alias{mMSNBC}
\title{Revised MSNBC Data}
\format{An object of class \code{list} of length 2.}
\usage{
data(mMSNBC)
}
\description{
This is a revised version of the MSNBC323 dataset in the R package ClickClust (Melnykov, 2016). This dataset contains the clickstreams without within-state transitions \code{x} and simulated time points \code{t}. See Gallaugher and McNicholas (2019) for further details.
}
\references{
Michael P.B. Gallaugher and Paul D. McNicholas (2019). Clustering and semi-supervised classification for clickstream data via mixture models. arXiv preprint arXiv:1802.04849v2.

Volodymyr Melnykov (2016). ClickClust: An R Package for Model-Based Clustering of Categorical Sequences. \emph{Journal of Statistical Software} \bold{74}(9), 1-34.
}
\keyword{data}