% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_Functions.R
\docType{data}
\name{SimData}
\alias{SimData}
\title{Simulated Data}
\format{An object of class \code{list} of length 2.}
\usage{
data(SimData)
}
\description{
This is a simulated dataset with 2 groups. It is in the form of a list with the first element being the list of states and the second element being the list of time stamps.
}
\keyword{data}
